tests/wpt/tests/css/css-text/parsing/text-wrap-mode-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-wrap-mode with valid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap-mode">
<meta name="assert" content="text-wrap-mode supports the full grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-wrap-mode", "wrap");
test_valid_value("text-wrap-mode", "nowrap");

test_valid_value("text-wrap-mode", "initial");
test_valid_value("text-wrap-mode", "inherit");
test_valid_value("text-wrap-mode", "unset");
test_valid_value("text-wrap-mode", "revert");
test_valid_value("text-wrap-mode", "revert-layer");
</script>
</body>
</html>
